Add test to check the conversion from subjects to results
