+++ src/renderer/src/stories/pages/guided-mode/setup/GuidedSubjects.js
@@ -5,6 +5,9 @@ import { validateOnChange } from "../../../../validation/index.js";
 import { SimpleTable } from "../../../SimpleTable.js";
 import { Table } from "../../../Table.js";
 
+import {convertSubjectsToResults} from './utils'
+
+
 export class GuidedSubjectsPage extends Page {
     constructor(...args) {
         super(...args);
@@ -13,8 +16,6 @@ export class GuidedSubjectsPage extends Page {
     footer = {
         onNext: () => {
             const { results, subjects } = this.info.globalState;
-
-            console.log("Results", subjects, results);
 
             try {
                 this.table.validate();
@@ -35,31 +36,7 @@ export class GuidedSubjectsPage extends Page {
                 return acc;
             }, {});
 
-            const toRemove = Object.keys(results).filter((sub) => !Object.keys(subjects).includes(sub));
-            for (let sub of toRemove) delete results[sub]; // Delete extra subjects from results
-
-            for (let subject in subjects) {
-                const { sessions = [] } = subjects[subject];
-                let subObj = results[subject];
-
-                if (!subObj) subObj = results[subject] = {};
-                else {
-                    const toRemove = Object.keys(subObj).filter((s) => !sessions.includes(s));
-                    for (let s of toRemove) delete subObj[s]; // Delete extra sessions from results
-                    if (!sessions.length && !Object.keys(subObj).length) delete results[subject]; // Delete subjects without sessions
-                }
-
-                for (let session of sessions) {
-                    if (!(session in subObj))
-                        subObj[session] = {
-                            source_data: { ...sourceDataObject },
-                            metadata: {
-                                NWBFile: { session_id: session },
-                                Subject: { subject_id: subject },
-                            },
-                        };
-                }
-            }
+            convertSubjectsToResults(results, subjects, sourceDataObject) // NOTE: This directly mutates the results object
 
             this.onTransition(1);
         },
