+++ tests/metadata.test.ts
@@ -8,6 +8,7 @@ import { createMockGlobalState } from './utils'
 
 import { Validator } from 'jsonschema'
 import { textToArray } from '../src/renderer/src/stories/forms/utils'
+import { convertSubjectsToResults } from '../src/renderer/src/stories/pages/guided-mode/setup/utils'
 
 var v = new Validator();
 
@@ -26,4 +27,15 @@ test('empty rows are not kept for strings converted to arrays', () => {
     expect(textToArray(' v1\n\n   v2 ')).toEqual(['v1', 'v2'])
     expect(textToArray(' v1\n \n   v2 ')).toEqual(['v1', 'v2'])
     expect(textToArray(' v1\n v3\n   v2 ')).toEqual(['v1', 'v3', 'v2'])
+})
+
+test('removing all existing sessions will maintain the related subject entry on the results object', () => {
+    const results = { subject: { original: {} } }
+
+    const copy = JSON.parse(JSON.stringify(results))
+
+    const subjects = { subject: { sessions: ['renamed'] } }
+
+    convertSubjectsToResults(results, subjects)
+    expect(Object.keys(results)).toEqual(Object.keys(copy))
 })
